Split ragger ethereum client into multiple smaller files
